+++ pages/pickpickpick/[id]/index.page.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react';
 
-import Image from 'next/image';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 
@@ -11,9 +10,8 @@ import { useModalStore } from '@stores/modalStore';
 
 import useIsMobile from '@hooks/useIsMobile';
 
+import MobileToListButton from '@components/common/mobileToListBtn/mobileToListButton';
 import MoreButton from '@components/common/moreButton';
-
-import listDots from '@public/image/list-dots.svg';
 
 import { ROUTES } from '@/constants/routes';
 
@@ -35,29 +33,9 @@ export default function Index() {
   const { selected, setSelected } = useSelectedStore();
   const isMobile = useIsMobile();
 
-  const [showBottom, setShowBottom] = useState(true);
-
   useEffect(() => {
     !isModalOpen && setSelected('신고 사유 선택');
   }, [isModalOpen]);
-
-  const handleScrollEvent = () => {
-    if (window.scrollY === 0) {
-      setShowBottom(true);
-      return;
-    }
-
-    setShowBottom(false);
-  };
-
-  const handleClickEvent = () => {
-    setShowBottom(true);
-  };
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScrollEvent);
-    window.addEventListener('click', handleClickEvent);
-  }, []);
 
   const { data: similarPicks } = useGetSimilarPick(id as string);
 
@@ -148,18 +126,7 @@ export default function Index() {
           </div>
         </div>
 
-        {isMobile && showBottom && (
-          <div className='h-[6.4rem]'>
-            <div className={`fixed left-0 right-0 bottom-0 px-[1.6rem] py-[1.9rem] bg-gray1 flex}`}>
-              <Link href={ROUTES.PICKPICKPICK.MAIN}>
-                <button className='st2 text-gray5 flex gap-[1rem] justify-center'>
-                  <Image src={listDots} alt='목록 아이콘' />
-                  목록으로
-                </button>
-              </Link>
-            </div>
-          </div>
-        )}
+        {isMobile && <MobileToListButton route={ROUTES.PICKPICKPICK.MAIN} />}
       </div>
 
       {isModalOpen && (
